# Fix Append modifying a possibly shared array