It's possible to have a multiError that is appended to multiple times
causing an underlying array to be shared and end up getting modified.

For now, remove the special case for `append(multiError, err)` which can
cause this bug.