@prashantv 2017-03-27
Just as a note, this does make performance significantly worse when we're appending errors to a `multierr`, which we should fix separately:

Before this change:
```
BenchmarkAppend/append_err_to_err/1-8           20000000                85.6 ns/op            64 B/op          2 allocs/op
BenchmarkAppend/append_err_to_err/2-8           10000000               220 ns/op             160 B/op          4 allocs/op
BenchmarkAppend/append_err_to_err/10-8           2000000               930 ns/op             800 B/op         14 allocs/op
BenchmarkAppend/append_multierr_to_multierr/1-8          5000000               289 ns/op             384 B/op          3 allocs/op
BenchmarkAppend/append_multierr_to_multierr/2-8          2000000               617 ns/op             928 B/op          6 allocs/op
BenchmarkAppend/append_multierr_to_multierr/10-8          300000              4880 ns/op           11680 B/op         30 allocs/op
```

After this change:
```
BenchmarkAppend/append_err_to_err/1-8           20000000                81.8 ns/op            64 B/op          2 allocs/op
BenchmarkAppend/append_err_to_err/2-8            5000000               281 ns/op             176 B/op          5 allocs/op
BenchmarkAppend/append_err_to_err/10-8           1000000              2109 ns/op            1648 B/op         29 allocs/op
BenchmarkAppend/append_multierr_to_multierr/1-8          5000000               282 ns/op             384 B/op          3 allocs/op
BenchmarkAppend/append_multierr_to_multierr/2-8          2000000               623 ns/op             928 B/op          6 allocs/op
BenchmarkAppend/append_multierr_to_multierr/10-8          300000              4826 ns/op           11680 B/op         30 allocs/op
```