@abhinav 2017-03-27
I haven't evaluated this idea fully but we may be able to solve this by
tracking the number of entries appended to the underlying array of lhs in
`Append(lhs, rhs)`.

```go
struct multiError struct {
  errors []error
  arrayLen *int
}

func appendToMultiError(merr multiErr, err error) multiError {
  merr.errors = append(merr.errors, err)
  *merr.arrayLen++
  // Original multiError now knows that its underlying array was appended to
  return merr
}
```

With that, the `Append` special case for `lhs.(multiError)` simply has to
check if `arrayLen > len(errors)` in which case it can't append.